Split 'pyhat_m3_examples' notebook into two versions.
'pyhat_m3_examples_tif.ipynb' Is a notebook that utilizes PyHAT to output and analyze .tif files.

'pyhat_m3_examples_img.ipynb' Is the sister notebook that does the same, but with .IMG files.

Since there may be an issue with the clip_ROI function, the .IMG notebook first analyzes the entire image, then outputs a human-readable subset of that image. 
Once clip_ROI can be implemented, this should greatly reduce computation time.

The version that processes .IMG files should be prioritized from here on out.